+++ Threeyes/SDK/Scripts/Component/Cursor/AC_AliveCursor.cs
@@ -3,7 +3,8 @@ using UnityEngine;
 /// <summary>
 /// The Root Manager for mod item
 /// </summary>
-public class AC_AliveCursor : MonoBehaviour, IAC_CommonSetting_IsAliveCursorActiveHandler
+public class AC_AliveCursor : MonoBehaviour
+	, IAC_CommonSetting_IsAliveCursorActiveHandler
 {
 	public static AC_AliveCursor Instance;
 	public virtual void Init()
@@ -11,10 +12,12 @@ public class AC_AliveCursor : MonoBehaviour, IAC_CommonSetting_IsAliveCursorActi
 		Instance = this;//Update Instance
 	}
 
+	#region Callback
 	public void OnIsAliveCursorActiveChanged(bool isActive)
 	{
 		SetActive(isActive);//显隐AC
 	}
+	#endregion
 
 	public void SetActive(bool isActive)
 	{